refactor: Replace all float types with double in proto definitions
- Changed all float field types to double for better precision
- Updated buf.validate.field annotations from .float to .double
- Regenerated all language bindings (C, C++, Go, Python, TypeScript, Rust, Java)
- This improves numerical precision and avoids float/double conversion issues

+++ output/rust/cmd.compass.rs
@@ -50,18 +50,18 @@ pub struct CalibrateCencel {}
 pub struct GetMeteo {}
 #[derive(Clone, Copy, PartialEq, ::prost::Message)]
 pub struct SetMagneticDeclination {
-    #[prost(float, tag = "1")]
-    pub value: f32,
+    #[prost(double, tag = "1")]
+    pub value: f64,
 }
 #[derive(Clone, Copy, PartialEq, ::prost::Message)]
 pub struct SetOffsetAngleAzimuth {
-    #[prost(float, tag = "1")]
-    pub value: f32,
+    #[prost(double, tag = "1")]
+    pub value: f64,
 }
 #[derive(Clone, Copy, PartialEq, ::prost::Message)]
 pub struct SetOffsetAngleElevation {
-    #[prost(float, tag = "1")]
-    pub value: f32,
+    #[prost(double, tag = "1")]
+    pub value: f64,
 }
 #[derive(Clone, Copy, PartialEq, ::prost::Message)]
 pub struct SetUseRotaryPosition {
